Move format module into protocol
