+++ src/protocol/packet.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::format;
+use crate::protocol::format;
 
 state_packets!(
     handshake Handshaking {
