+++ src/types/metadata.rs
@@ -19,7 +19,7 @@ use std::fmt;
 use crate::protocol;
 use crate::protocol::Serializable;
 use crate::protocol::LenPrefixed;
-use crate::format;
+use crate::protocol::format;
 use crate::item;
 use crate::shared::Position;
 use crate::protocol::nbt;
